refactor: replace var with const
